^Y.
Implemented restore item functionality.
Added Restore Deleted Items link to index page. Reused filter view bassing deleted=true to list deleted items.
An item can be selected with the existing list implementation to redirect to the item page.
Modified the item page to check whether the item is deleted and if so show a Restore button and Hide Delete and Update buttons.
Re-used update-item passing deleted = false when the Restore button is clicked and gave success message in HTML.